add suggestions; refactor jsonl.dump;
