+++ src/gluonts/dataset/repository/datasets.py
@@ -23,7 +23,7 @@ from gluonts.dataset.repository._gp_copula_2019 import (
     generate_gp_copula_dataset,
 )
 from gluonts.dataset.repository._lstnet import generate_lstnet_dataset
-from gluonts.dataset.repository._zenodo import generate_forecasting_dataset
+from gluonts.dataset.repository._tsf_datasets import generate_forecasting_dataset
 from gluonts.dataset.repository._m3 import generate_m3_dataset
 from gluonts.dataset.repository._m4 import generate_m4_dataset
 from gluonts.dataset.repository._m5 import generate_m5_dataset
